pyproject.toml
[tool.poetry]
name = "portman"
version = "0.1.0"
description = ""
authors = ["sieginglion <sieginglion@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiofiles = "^23.1.0"
arrow = "^1.2.3"
fastapi = "^0.100.0"
httpx = "^0.24.1"
numba = "^0.57.1"
python-dotenv = "^1.0.0"
uvicorn = "^0.23.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
